Add files via upload
This design file made use of tidy3d autograd. The previous versions were using Jax which is deprecated since tidy3d 2.7.0. 
The structure is also more serious than previous versions, which contains a bilayer (a metal layer plus a dielectric layer) in the substrate and five bilayers in the cuboid. 
A great upgrade.