Create design2_v3.1_variable_LR.ipynb
An incomplete version of optimization, that the loss function is purely related to the distance between the peak of the optimizing flux and the target. An update from design2_v3.0 is at the peak estimation function find_peak_distance_from_flux_2ndD. The last version uses find_peak_distance_from_flux and the estimation is poor on smooth spectrum. The second derivative estimation is robust.